requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.0.8
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cx-Freeze==6.12.0
cycler==0.12.1
debugpy==1.6.3
decorator==5.1.1
emoji==2.10.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==1.1.1
fastapi==0.115.0
fonttools==4.54.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
idna==3.10
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-core==4.11.1
jupyter_client==7.4.2
kiwisolver==1.4.7
lief==0.12.2
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
MouseInfo==0.1.3
narwhals==1.9.2
nest-asyncio==1.5.6
numpy==2.1.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.0.10
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parso==0.8.3
pickleshare==0.7.5
pillow==10.4.0
playwright==1.44.0
plotly==5.24.1
prompt-toolkit==3.0.31
protobuf==5.28.2
psutil==5.9.2
psycopg2==2.9.9
pure-eval==0.2.2
pyarrow==17.0.0
PyAutoGUI==0.9.53
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pyee==11.1.0
pygame==2.5.2
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyodbc==5.1.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.0.4
pytz==2024.2
pywin32==304
pyzmq==24.0.1
referencing==0.35.1
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
selenium==4.18.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.5.1
starlette==0.38.2
streamlit==1.39.0
tenacity==9.0.0
tinydb==4.7.0
toml==0.10.2
tornado==6.4.1
traitlets==5.4.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.30.6
watchdog==5.0.3
wcwidth==0.2.5
webdriver-manager==4.0.1
wsproto==1.2.0
